requirements.txt
requests
pandas
google-generativeai
python-dotenv
schedule
beautifulsoup4
lxml
